refactor: domain 내부에 패키지를 정의해 각 모델을 적절한 위치로 이동
