+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,5 +1,7 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.cardholder.CardHolder;
 import java.util.List;
 
 public class Dealer {
