+++ src/main/java/blackjack/domain/card/CardRank.java
@@ -1,4 +1,4 @@
-package blackjack.common;
+package blackjack.domain.card;
 
 import java.util.List;
