+++ src/main/java/blackjack/domain/card/CardSuit.java
@@ -1,4 +1,4 @@
-package blackjack.common;
+package blackjack.domain.card;
 
 public enum CardSuit {
