+++ src/main/java/blackjack/domain/cardholder/CardHolder.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.cardholder;
 
+import blackjack.domain.card.Card;
 import java.util.List;
 
 public interface CardHolder {
