+++ src/main/java/blackjack/domain/cardholder/Hand.java
@@ -1,7 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.cardholder;
 
 import blackjack.common.Constants;
 import blackjack.common.ErrorMessage;
+import blackjack.domain.card.Card;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
