+++ src/main/java/blackjack/domain/result/PlayersResult.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
+import blackjack.domain.result.GameResultType;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.Map;
