+++ src/test/java/blackjack/domain/card/CardTest.java
@@ -1,7 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import blackjack.common.CardRank;
-import blackjack.common.CardSuit;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardRank;
+import blackjack.domain.card.CardSuit;
 import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
