+++ src/test/java/blackjack/domain/cardholder/HandTest.java
@@ -1,7 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.cardholder;
 
-import blackjack.common.CardRank;
-import blackjack.common.CardSuit;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardRank;
+import blackjack.domain.card.CardSuit;
+import blackjack.domain.cardholder.Hand;
 import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
